Rename transforms subdirectory as per upstream conventions.
