+++ mlir/include/mlir-extensions/Transforms/index_type_propagation.hpp
