+++ mlir/lib/Dialect/gpu_runtime/IR/gpu_runtime_ops.cpp
@@ -31,7 +31,7 @@
 #include <llvm/ADT/SmallBitVector.h>
 #include <llvm/ADT/TypeSwitch.h>
 
-#include "mlir-extensions/transforms/const_utils.hpp"
+#include "mlir-extensions/Transforms/const_utils.hpp"
 
 namespace MemoryEffects = ::mlir::MemoryEffects;
