+++ mlir/lib/Dialect/plier_util/dialect.cpp
@@ -36,7 +36,7 @@
 #include <llvm/ADT/SmallBitVector.h>
 #include <llvm/ADT/TypeSwitch.h>
 
-#include "mlir-extensions/transforms/const_utils.hpp"
+#include "mlir-extensions/Transforms/const_utils.hpp"
 
 namespace MemoryEffects = ::mlir::MemoryEffects;
