+++ mlir/lib/Transforms/canonicalize_reductions.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mlir-extensions/transforms/canonicalize_reductions.hpp"
+#include "mlir-extensions/Transforms/canonicalize_reductions.hpp"
 
 #include <mlir/Dialect/MemRef/IR/MemRef.h>
 #include <mlir/Dialect/SCF/SCF.h>
