+++ mlir/lib/Transforms/cast_lowering.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mlir-extensions/transforms/cast_lowering.hpp"
+#include "mlir-extensions/Transforms/cast_lowering.hpp"
 
 #include <mlir/Transforms/DialectConversion.h>
