+++ mlir/lib/Transforms/cast_utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mlir-extensions/transforms/cast_utils.hpp"
+#include "mlir-extensions/Transforms/cast_utils.hpp"
 #include "mlir-extensions/Dialect/plier_util/dialect.hpp"
 
 #include <mlir/Dialect/Arithmetic/IR/Arithmetic.h>
