+++ mlir/lib/Transforms/const_utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mlir-extensions/transforms/const_utils.hpp"
+#include "mlir-extensions/Transforms/const_utils.hpp"
 
 #include <mlir/IR/BuiltinTypes.h>
 #include <mlir/IR/OpDefinition.h>
