+++ mlir/lib/Transforms/func_utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mlir-extensions/transforms/func_utils.hpp"
+#include "mlir-extensions/Transforms/func_utils.hpp"
 
 #include <llvm/ADT/StringRef.h>
 #include <mlir/Dialect/Func/IR/FuncOps.h>
