+++ mlir/lib/Transforms/if_rewrites.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mlir-extensions/transforms/if_rewrites.hpp"
+#include "mlir-extensions/Transforms/if_rewrites.hpp"
 
 #include <mlir/Dialect/Arithmetic/IR/Arithmetic.h>
 #include <mlir/Dialect/SCF/SCF.h>
