+++ mlir/lib/Transforms/index_type_propagation.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mlir-extensions/transforms/index_type_propagation.hpp"
+#include "mlir-extensions/Transforms/index_type_propagation.hpp"
 
 #include <mlir/Dialect/Arithmetic/IR/Arithmetic.h>
 #include <mlir/IR/PatternMatch.h>
