+++ mlir/lib/Transforms/loop_utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "mlir-extensions/transforms/loop_utils.hpp"
+#include "mlir-extensions/Transforms/loop_utils.hpp"
 
 #include <llvm/ADT/SmallVector.h>
 
@@ -25,8 +25,8 @@
 #include <mlir/Support/LogicalResult.h>
 
 #include "mlir-extensions/Dialect/plier/dialect.hpp"
-#include "mlir-extensions/transforms/cast_utils.hpp"
-#include "mlir-extensions/transforms/const_utils.hpp"
+#include "mlir-extensions/Transforms/cast_utils.hpp"
+#include "mlir-extensions/Transforms/const_utils.hpp"
 
 namespace {
 template <typename Op>
